Performance updates
Fixed the vertical gap between chunks by changing the drawing mode for each one to shapeMode(CORNER). Optimised chunk rendering by only recalculatin the chunks in front of the player, depending on his movement direction in the method updateChunks. It's called only when new chunks have to be rendered(when the player pases a chunk threshold). because of the rendering technique for chunks(each row is created as it's own shape), a problem is present, where individual triangles can not be colored, only entire rows.