Water update
Added water pools. Water is drawn on the entire chunk and the height is modulated by a noise function. Performance does not visibly suffer, even though there are 2 double for loops in the water display function, because of the low level of vertecies. May need to change later, depending on the outcome. Slight bug when facing North-West where the player can see the edges of chunks in the water, because of the overlapping aplhas. Added a Open Simplex Implementation for performance testing later;