Multithreading
Added a thread execution per each chunk calculation in updateChunks(). This decreases the calculation time by 3 beacause of the parallel creation of chunks. Currently the end of updateChunks() waits for all threads to finish, whitch produces results of about 0.130 s per method execution. This can be averted by not waiting for the threads, but that produces some texturing bugs in the chunks.